requirements.txt
boto3==1.34.157
pyffmpeg==2.4.2
pillow==10.4.0
soundfile=0.12.1